Fetch fails from `WeatherRepository`
- Change data type of lat and lon

+++ dranzer-app/app/src/main/java/com/developers/dranzer/data/WeatherRepository.kt
@@ -5,7 +5,7 @@ import com.developers.dranzer.data.network.WeatherApi
 
 class WeatherRepository(private val weatherApi: WeatherApi) {
 
-    suspend fun getCurrentWeather(latitude: Float, longitude: Float): Result<CurrentWeatherData> {
+    suspend fun getCurrentWeather(latitude: Double, longitude: Double): Result<CurrentWeatherData> {
         return runCatching { weatherApi.getCurrentWeather(latitude, longitude) }
     }
 }
